refactor(frontend): 🧑‍💻 easier context imports
