+++ app/frontend/src/context/index.ts
@@ -1,0 +1,5 @@
+import { QueryClientProvider } from "./query-provider";
+import { RouterProvider } from "./router-provider";
+import { ThemeProvider } from "./theme-provider";
+
+export { QueryClientProvider, ThemeProvider, RouterProvider };
