+++ app/frontend/src/hooks/index.ts
@@ -1,0 +1,3 @@
+import { useTheme } from "./use-theme";
+
+export { useTheme };
